backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
email-validator==2.3.0
pytest==8.4.1
httpx==0.28.1
redis==5.0.1
psutil==5.9.6
